simplified katana runner
